Login fetches user courses information, page account selects and displays
