+++ src/state/session/selectors.ts
@@ -1,1 +1,4 @@
 export const getUser = state => state.session.user
+export const getCourses = state => state.session.courses
+export const getCompletedLessons = state => state.session.completedLessons
+export const getTodoLessons = state => state.session.todoLessons
